gulp clean / cache
took out gulp cache, b/c it doesn't work well with cleaning up files.
I'm not sure how to combine those two ideas: we want to clean out the
dist/ folder, so that any zombie files lurking around are removed, and
we'd like not to have to run already compiled ish. Oh well. wait, does
deleting a file count as a change to a file? hmm... anyway, works as it
is now.